refactor: use delta table
